refactor timestamp event log helpers to shared helpers
