+++ src/lib/event-logs.ts
@@ -1,5 +1,3 @@
-import { format, formatDistanceToNowStrict } from "date-fns"
-
 import { type DiffEntry } from "@/components/diff"
 
 const DESTRUCTIVE_EVENT_RE = [
@@ -32,85 +30,6 @@ export function eventLogBadgeVariant(
   }
 
   return "secondary"
-}
-
-export function formatEventLogRelativeTime(value: string): string {
-  const date = new Date(value)
-  if (Number.isNaN(date.getTime())) return value
-
-  return formatDistanceToNowStrict(date, { addSuffix: true })
-}
-
-export function formatEventLogTimestamp(value: string): string {
-  const date = new Date(value)
-  if (Number.isNaN(date.getTime())) return value
-
-  return format(date, "PP p")
-}
-
-export type EventLogZonedTimestamp = {
-  label: string
-  date: string
-  time: string
-}
-
-function formatEventLogZonedTimestamp(
-  value: string,
-  timeZone: string,
-  label: string,
-): EventLogZonedTimestamp {
-  const date = new Date(value)
-  if (Number.isNaN(date.getTime())) {
-    return {
-      label,
-      date: value,
-      time: "",
-    }
-  }
-
-  const parts = new Intl.DateTimeFormat("en-US", {
-    timeZone,
-    month: "long",
-    day: "numeric",
-    year: "numeric",
-    hour: "numeric",
-    minute: "2-digit",
-    second: "2-digit",
-    hour12: true,
-  }).formatToParts(date)
-
-  const valueFor = (type: Intl.DateTimeFormatPartTypes) =>
-    parts.find((part) => part.type === type)?.value ?? ""
-
-  return {
-    label,
-    date: `${valueFor("month")} ${valueFor("day")}, ${valueFor("year")}`,
-    time:
-      [valueFor("hour"), valueFor("minute"), valueFor("second")].join(":") +
-      ` ${valueFor("dayPeriod")}`,
-  }
-}
-
-export function formatEventLogUtcTimestamp(
-  value: string,
-): EventLogZonedTimestamp {
-  return formatEventLogZonedTimestamp(value, "UTC", "UTC")
-}
-
-export function formatEventLogLocalTimestamp(
-  value: string,
-): EventLogZonedTimestamp {
-  const timeZone = Intl.DateTimeFormat().resolvedOptions().timeZone
-  const date = new Date(value)
-  if (Number.isNaN(date.getTime())) {
-    return {
-      label: "Local",
-      date: value,
-      time: "",
-    }
-  }
-
-  return formatEventLogZonedTimestamp(value, timeZone, "Local")
 }
 
 export function visibleMetadata(
